requirements.txt
dnspython==2.8.0
et_xmlfile==2.0.0
greenlet==3.3.2
openpyxl==3.1.5
psycopg2-binary==2.9.11
pymongo==4.16.0
PyMySQL==1.1.2
python-dotenv==1.2.2
SQLAlchemy==2.0.48
typing_extensions==4.15.0
